+++ src/convert/types.js
@@ -6,7 +6,6 @@ const MEDIA_TYPES = [
   'interactive',
   'location',
   'placeholder',
-  'quote',
   'share',
   'video'
 ]
@@ -21,6 +20,7 @@ const HTML_TYPES = [
   'list',
   'ol',
   'p',
+  'quote',
   'text',
   'ul'
 ]