+++ test/index.js
@@ -1,5 +1,9 @@
+import chai from 'chai'
+chai.config.truncateThreshold = 0
+
 import './ed'
 import './components/image'
-import './plugins/tap-add-text'
+import './convert/doc-to-grid.js'
+import './convert/grid-to-doc.js'
 import './plugins/widget'
 import './schema/block-meta'